refactor(player): relocate leaf transport views out of VLCStreamPresenter (M3)
Moved the 3 self-contained UIKit leaf views (TVScrubBar, ChapterChip,
PassthroughView) to PlayerTransportViews.swift. They have zero coupling to
the view controller's playback/HUD state (own closures/props only), so this
is a pure relocation: VLCStreamPresenter.swift 1768 → 1659 LOC, no behavior
change.

Deeper HUD/sleep-timer/skip controller extraction stays deferred — it is
interwoven with engine pause/play + the tick loop and is only validatable
on real hardware (see Audit_post_vlc.md).

iOS+tvOS build + 40 tests green. xcodegen regenerated.
